@@ -10,7 +10,7 @@
     "email": "pleerock.me@gmail.com"
   },
   "engines": {
-    "node": ">= 12.9.0"
+    "node": ">=16.13.0"
   },
   "exports": {
     ".": {
@@ -93,10 +93,13 @@
     "cloud-spanner-orm"
   ],
   "devDependencies": {
+    "@tsconfig/node16": "^16.1.1",
     "@types/app-root-path": "^1.2.4",
     "@types/chai": "^4.3.4",
     "@types/chai-as-promised": "^7.1.5",
     "@types/debug": "^4.1.7",
+    "@types/gulp-rename": "^2.0.6",
+    "@types/gulp-sourcemaps": "^0.0.38",
     "@types/mkdirp": "^1.0.2",
     "@types/mocha": "^10.0.1",
     "@types/node": "^18.13.0",
@@ -105,7 +108,7 @@
     "@types/source-map-support": "^0.5.6",
     "@types/uuid": "^9.0.0",
     "@types/yargs": "^17.0.22",
-    "@typescript-eslint/eslint-plugin": "^6.0.0",
+    "@typescript-eslint/eslint-plugin": "^6.17.0",
     "better-sqlite3": "^8.1.0",
     "chai": "^4.3.7",
     "chai-as-promised": "^7.1.1",
@@ -140,9 +143,9 @@
     "source-map-support": "^0.5.21",
     "sql.js": "^1.8.0",
     "sqlite3": "^5.1.4",
-    "ts-node": "^10.9.1",
+    "ts-node": "^10.9.2",
     "typeorm-aurora-data-api-driver": "^2.4.4",
-    "typescript": "^4.9.5"
+    "typescript": "^5.3.3"
   },
   "peerDependencies": {
     "@google-cloud/spanner": "^5.18.0",
@@ -227,7 +230,7 @@
     "dotenv": "^16.0.3",
     "glob": "^10.3.10",
     "mkdirp": "^2.1.3",
-    "reflect-metadata": "^0.1.13",
+    "reflect-metadata": "^0.2.1",
     "sha.js": "^2.4.11",
     "tslib": "^2.5.0",
     "uuid": "^9.0.0",
