@@ -11,6 +11,7 @@ import {
     OptimisticLockCanNotBeUsedError,
     OptimisticLockVersionMismatchError,
     PessimisticLockTransactionRequiredError,
+    QueryRunner,
 } from "../../../../src"
 import { PostWithVersion } from "./entity/PostWithVersion"
 import { expect } from "chai"
@@ -135,7 +136,7 @@ describe("repository > find options > locking", () => {
                     const originalQuery = entityManager.queryRunner!.query.bind(
                         entityManager.queryRunner,
                     )
-                    entityManager.queryRunner!.query = (...args: any[]) => {
+                    entityManager.queryRunner!.query = (...args: Parameters<QueryRunner['query']>) => {
                         executedSql.push(args[0])
                         return originalQuery(...args)
                     }
@@ -175,7 +176,7 @@ describe("repository > find options > locking", () => {
                     const originalQuery = entityManager.queryRunner!.query.bind(
                         entityManager.queryRunner,
                     )
-                    entityManager.queryRunner!.query = (...args: any[]) => {
+                    entityManager.queryRunner!.query = (...args: Parameters<QueryRunner['query']>) => {
                         executedSql.push(args[0])
                         return originalQuery(...args)
                     }
@@ -204,7 +205,7 @@ describe("repository > find options > locking", () => {
                     const originalQuery = entityManager.queryRunner!.query.bind(
                         entityManager.queryRunner,
                     )
-                    entityManager.queryRunner!.query = (...args: any[]) => {
+                    entityManager.queryRunner!.query = (...args: Parameters<QueryRunner['query']>) => {
                         executedSql.push(args[0])
                         return originalQuery(...args)
                     }
@@ -243,7 +244,7 @@ describe("repository > find options > locking", () => {
                     const originalQuery = entityManager.queryRunner!.query.bind(
                         entityManager.queryRunner,
                     )
-                    entityManager.queryRunner!.query = (...args: any[]) => {
+                    entityManager.queryRunner!.query = (...args: Parameters<QueryRunner['query']>) => {
                         executedSql.push(args[0])
                         return originalQuery(...args)
                     }
@@ -285,7 +286,7 @@ describe("repository > find options > locking", () => {
                     const originalQuery = entityManager.queryRunner!.query.bind(
                         entityManager.queryRunner,
                     )
-                    entityManager.queryRunner!.query = (...args: any[]) => {
+                    entityManager.queryRunner!.query = (...args: Parameters<QueryRunner['query']>) => {
                         executedSql.push(args[0])
                         return originalQuery(...args)
                     }
@@ -322,7 +323,7 @@ describe("repository > find options > locking", () => {
                     const originalQuery = entityManager.queryRunner!.query.bind(
                         entityManager.queryRunner,
                     )
-                    entityManager.queryRunner!.query = (...args: any[]) => {
+                    entityManager.queryRunner!.query = (...args: Parameters<QueryRunner['query']>) => {
                         executedSql.push(args[0])
                         return originalQuery(...args)
                     }
@@ -360,7 +361,7 @@ describe("repository > find options > locking", () => {
                     const originalQuery = entityManager.queryRunner!.query.bind(
                         entityManager.queryRunner,
                     )
-                    entityManager.queryRunner!.query = (...args: any[]) => {
+                    entityManager.queryRunner!.query = (...args: Parameters<QueryRunner['query']>) => {
                         executedSql.push(args[0])
                         return originalQuery(...args)
                     }
