@@ -21,9 +21,9 @@ const wrap = (
 
         return new Promise<void>((ok, fail) => {
             if (fn.length > 1) {
-                fn.call(context, (err: any) => (err ? fail(err) : ok()))
+                (fn as Func).call(context as unknown as Context, (err: any) => (err ? fail(err) : ok()))
             } else {
-                ok(fn.call(context))
+                ok((fn as AsyncFunc).call(context as unknown as Context))
             }
         }).then(
             (e: any) => assert.fail("Expected this test to fail"),
